fix: adjust imports to avoid circular importing
