+++ dis_snek/models/snek/converters.py
@@ -1,6 +1,5 @@
 import re
 import typing
-import inspect
 from typing import TypeVar, Protocol, Any, Optional, List, Callable
 
 from dis_snek.client.errors import Forbidden, HTTPException
@@ -33,13 +32,13 @@ from dis_snek.models.discord.channel import (
     TYPE_VOICE_CHANNEL,
     TYPE_MESSAGEABLE_CHANNEL,
 )
+from dis_snek.models.snek.protocols import Converter
 from dis_snek.models.snek.context import Context
 from dis_snek.client.utils.misc_utils import get_object_name, get_parameters
 from dis_snek.client.errors import BadArgument
 
 
 __all__ = (
-    "Converter",
     "LiteralConverter",
     "IDConverter",
     "SnowflakeConverter",
@@ -72,12 +71,6 @@ __all__ = (
 
 T = TypeVar("T")
 T_co = TypeVar("T_co", covariant=True)
-
-
-@typing.runtime_checkable
-class Converter(Protocol[T_co]):
-    async def convert(self, ctx: Context, argument: Any) -> T_co:
-        raise NotImplementedError("Derived classes need to implement this.")
 
 
 class LiteralConverter(Converter):
@@ -421,20 +414,3 @@ SNEK_MODEL_TO_CONVERTER: dict[type, type[Converter]] = {
     PartialEmoji: PartialEmojiConverter,
     CustomEmoji: CustomEmojiConverter,
 }
-
-
-def _get_converter_function(anno: type[Converter] | Converter, name: str) -> Callable[[Context, str], Any]:
-    num_params = len(get_parameters(anno.convert))
-
-    # if we have three parameters for the function, it's likely it has a self parameter
-    # so we need to get rid of it by initing - typehinting hates this, btw!
-    # the below line will error out if we aren't supposed to init it, so that works out
-    actual_anno: Converter = anno() if num_params == 3 else anno  # type: ignore
-    # we can only get to this point while having three params if we successfully inited
-    if num_params == 3:
-        num_params -= 1
-
-    if num_params != 2:
-        ValueError(f"{get_object_name(anno)} for {name} is invalid: converters must have exactly 2 arguments.")
-
-    return actual_anno.convert
